Remove Slaq::IO
