--- lib/slaq/io/slack/messege.rb
@@ -1,8 +1,0 @@
-module Slaq
-  module IO
-    class Slack
-      module Message
-      end
-    end
-  end
-end
